requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
pgvector>=0.3.0

# LLM providers
openai>=1.50.0
anthropic>=0.36.0

# Embeddings & retrieval
rank-bm25>=0.2.2
sentence-transformers>=3.0.0

# HTTP client
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0

# Dev tooling
rich>=13.0.0
